broke the login page but made some more progress
