+++ app/src/data/DataMenu.module.css
@@ -1,0 +1,19 @@
+.container {
+  display: flex;
+  flex-direction: column;
+  width: 100%;
+  height: 100%;
+  align-items: center;
+  justify-content: center;
+}
+
+.menuGroup {
+  display: flex;
+  flex-direction: column;
+  gap: 16px;
+}
+
+.menuButton {
+  background-color: var(--color-primary);
+  color: var(--color-primary-contrastText);
+}
