+++ app/src/index.css
@@ -10,10 +10,15 @@
   --text-h1-size: 5rem;
   --text-h2-size: 2rem;
   --text-body-size: 1.5em;
+
+  --grid-border-margin-width: 20px;
+  --grid-border-border-width: 5px;
 }
 
 * {
   box-sizing: border-box;
+  margin: 0;
+  padding: 0;
 }
 
 html,
